Updated script to support namespacing of Open Graph protocol.
The old version of the script was limited to only parsing for the og
namespace. This revision adds support for parsing for any namespaces
desired, which unfortunately breaks the expected response for prior
implementations however it is a necessary improvement.

As a note the old implementation failed to work on sites following
current Open Graph standards from http://ogp.me due to modifications
in the namespace prefix declaration. This revision forgoes the forcing
of prefix detection from jsdom since the Open Graph standard has changed
at least three times that I am aware of.